Add backward compatbility for RecordSerializer and RecordDeserializer (#5052)
* Add backward compatbility for RecordSerializer and RecordDeserializer

* fix circular import

* fix test

---------
