+++ src/sagemaker/base_serializers.py
@@ -22,7 +22,7 @@ import numpy as np
 from pandas import DataFrame
 from six import with_metaclass
 
-from sagemaker.amazon.common import write_numpy_to_dense_tensor
+from sagemaker.serializer_utils import write_numpy_to_dense_tensor
 from sagemaker.utils import DeferredError
 
 try:
